refactor: 내가 작성한 글 없을 때 UI 수정
- 메세지 상수화

+++ frontend/src/pages/MyPagePosts/index.js
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import { css } from '@emotion/react';
 import { useHistory } from 'react-router-dom';
-import { PATH } from '../../constants';
+import { ALERT_MESSAGE, CONFIRM_MESSAGE, PATH } from '../../constants';
 import { Button, BUTTON_SIZE } from '../../components';
 import { requestGetMyPosts } from '../../service/requests';
-import { Content, Description, Mission, Title, Tags, PostItem, ButtonList } from './styles';
+import { Content, Description, Mission, Title, Tags, PostItem, ButtonList, NoPost } from './styles';
 import { useSelector } from 'react-redux';
 import usePost from '../../hooks/usePost';
 
@@ -60,11 +60,13 @@ const MyPagePosts = () => {
     }
   };
 
-  const deleteTargetPost = async (id) => {
+  const onDeletePost = async (id) => {
+    if (!window.confirm(CONFIRM_MESSAGE.DELETE_POST)) return;
+
     await deletePost(id, accessToken);
 
     if (postError) {
-      alert('글을 삭제할 수 없습니다.');
+      alert(ALERT_MESSAGE.FAIL_TO_DELETE_POST);
       return;
     }
 
@@ -77,55 +79,57 @@ const MyPagePosts = () => {
 
   return (
     <>
-      {posts.length
-        ? posts.map((post) => {
-            const { id, mission, title, tags } = post;
+      {posts.length ? (
+        posts.map((post) => {
+          const { id, mission, title, tags } = post;
 
-            return (
-              <PostItem
-                key={id}
-                size="SMALL"
-                onClick={goTargetPost(id)}
-                onMouseEnter={() => setHoveredPostId(id)}
-                onMouseLeave={() => setHoveredPostId(0)}
-              >
-                <Content>
-                  <Description>
-                    <Mission>{mission.name}</Mission>
-                    <Title>{title}</Title>
-                    <Tags>
-                      {tags.map(({ id, name }) => (
-                        <span key={id}>{`#${name} `}</span>
-                      ))}
-                    </Tags>
-                  </Description>
-                </Content>
-                {hoverdPostId === id && (
-                  <ButtonList>
-                    <Button
-                      size={BUTTON_SIZE.X_SMALL}
-                      type="button"
-                      css={EditButtonStyle}
-                      alt="수정 버튼"
-                      onClick={() => goEditTargetPost(id)}
-                    >
-                      수정
-                    </Button>
-                    <Button
-                      size={BUTTON_SIZE.X_SMALL}
-                      type="button"
-                      css={DeleteButtonStyle}
-                      alt="삭제 버튼"
-                      onClick={() => deleteTargetPost(id)}
-                    >
-                      삭제
-                    </Button>
-                  </ButtonList>
-                )}
-              </PostItem>
-            );
-          })
-        : '작성한 글이 없습니다.'}
+          return (
+            <PostItem
+              key={id}
+              size="SMALL"
+              onClick={goTargetPost(id)}
+              onMouseEnter={() => setHoveredPostId(id)}
+              onMouseLeave={() => setHoveredPostId(0)}
+            >
+              <Content>
+                <Description>
+                  <Mission>{mission.name}</Mission>
+                  <Title>{title}</Title>
+                  <Tags>
+                    {tags.map(({ id, name }) => (
+                      <span key={id}>{`#${name} `}</span>
+                    ))}
+                  </Tags>
+                </Description>
+              </Content>
+              {hoverdPostId === id && (
+                <ButtonList>
+                  <Button
+                    size={BUTTON_SIZE.X_SMALL}
+                    type="button"
+                    css={EditButtonStyle}
+                    alt="수정 버튼"
+                    onClick={() => goEditTargetPost(id)}
+                  >
+                    수정
+                  </Button>
+                  <Button
+                    size={BUTTON_SIZE.X_SMALL}
+                    type="button"
+                    css={DeleteButtonStyle}
+                    alt="삭제 버튼"
+                    onClick={() => onDeletePost(id)}
+                  >
+                    삭제
+                  </Button>
+                </ButtonList>
+              )}
+            </PostItem>
+          );
+        })
+      ) : (
+        <NoPost>작성한 글이 없습니다 🥲</NoPost>
+      )}
     </>
   );
 };
